#393 feat: Create root component and route for Glossary, remove GlossaryConsumables mock component
