--- src/pages/Glossary/components/GlossaryConsumables/GlossaryConsumables.tsx
@@ -1,5 +1,0 @@
-export function GlossaryConsumables() {
-    return (
-        <div>Glossary Consumables</div>
-    );
-}
