+++ src/pages/Glossary/components/GlossaryWearables/styles.ts
@@ -2,6 +2,6 @@ import { createStyles, makeStyles } from '@mui/styles';
 
 export const styles = makeStyles(() => createStyles({
     glossaryWearablesContainer: {
-        padding: '8px 24px'
+        padding: '56px 16px 12px'
     }
 }));
