+++ src/pages/Glossary/components/index.ts
@@ -1,0 +1,2 @@
+export { GlossaryRoot } from './GlossaryRoot/GlossaryRoot';
+export { GlossaryWearables } from '../components/GlossaryWearables/GlossaryWearables';
